bug fix
bug fix: The terminal command to activate the virtual environment is wrong. The assigned virtual environment folder name is ".django_venv/bin/activate" therefore this is the name that should be used when activating the virtual environment instead of ".venv/bin/activate". This fixed the bug I was having, and I assume others may have the same issue.